perf: replace xxhash-wasm with @node-rs/xxhash
Replace xxhash-wasm with @node-rs/xxhash.

The latter has faster throughput and factor less call overhead. xxhash
invocations are turned into trivial synchronous calls.
