+++ src/hashline.ts
@@ -5,7 +5,7 @@
  * Key additions ported: merge detection, confusable hyphens, restoreOldWrappedLines.
  */
 
-import xxhashWasm from "xxhash-wasm";
+import { xxh32 } from "@node-rs/xxhash";
 import { throwIfAborted } from "./runtime";
 import type { PtcLine } from "./ptc-value.js";
 
@@ -65,28 +65,16 @@ const CONFUSABLE_HYPHENS_RE = /[\u2010\u2011\u2012\u2013\u2014\u2212\uFE63\uFF0D
 const HASH_RELOCATION_WINDOW_BASE = 20;
 const HASH_RELOCATION_WINDOW_CAP = 100;
 
-let h32Fn: ((input: string, seed?: number) => number) | null = null;
-let initPromise: Promise<void> | null = null;
 
+/** @deprecated xxh32 is synchronous with @node-rs/xxhash. */
 export async function ensureHashInit(): Promise<void> {
-	if (h32Fn) return;
-	if (!initPromise) {
-		initPromise = xxhashWasm().then((hasher) => {
-			h32Fn = hasher.h32;
-		});
-	}
-	await initPromise;
-}
-
-function xxh32(input: string): number {
-	if (!h32Fn) throw new Error("Hash not initialized — call ensureHashInit() first");
-	return h32Fn(input, 0) >>> 0;
+	// no-op
 }
 
 export function computeLineHash(_idx: number, line: string): string {
 	if (line.endsWith("\r")) line = line.slice(0, -1);
 	line = line.replace(/\s+/g, "");
-	return DICT[xxh32(line) % HASH_MOD];
+	return DICT[(xxh32(line, 0) >>> 0) % HASH_MOD];
 }
 
 const DISPLAY_CONTROL_CHAR_RE = /[\x00-\x08\x0b\x0c\x0e-\x1f]/g;
